+++ src/formatted-input-currency.component.jsx
@@ -61,7 +61,15 @@ class FormattedInputCurrency extends React.PureComponent {
       ? val.length - decimalSeparatorIndex - 1
       : 0;
 
-    const value = formatCurrencyAmount(val, { decimals, thousandSeparator, decimalSeparator });
+    let value = val;
+    // formatCurrencyAmount input must be either number or number string
+    if (thousandSeparator) {
+      value = val.replace(new RegExp(`\\${thousandSeparator}`, 'g'), '');
+    }
+    if (decimalSeparator !== '.') {
+      value = val.replace(new RegExp(`\\${decimalSeparator}`, 'g'), '.');
+    }
+    value = formatCurrencyAmount(value, { decimals, thousandSeparator, decimalSeparator });
 
     // this avoids minus sign removal in the editing phase like this: 0.00 --> -0.00 --> -10.00
     if (val === `-${value}`) return val;